move mysql repo to repo sub folder
