+++ internal/app/repository.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	"github.com/go-seidon/chariot/internal/repository"
-	repository_mysql "github.com/go-seidon/chariot/internal/repository-mysql"
-	"github.com/go-seidon/provider/mysql"
+	"github.com/go-seidon/chariot/internal/repository/mysql"
+	mysql_client "github.com/go-seidon/provider/mysql"
 	gorm_mysql "gorm.io/driver/mysql"
 	"gorm.io/gorm"
 	"gorm.io/plugin/dbresolver"
@@ -22,21 +22,21 @@ func NewDefaultRepository(config *Config) (repository.Provider, error) {
 
 	var repo repository.Provider
 	if config.RepositoryProvider == repository.PROVIDER_MYSQL {
-		dbPrimary, err := mysql.NewClient(
-			mysql.WithAuth(config.MySQLPrimaryUser, config.MySQLPrimaryPassword),
-			mysql.WithConfig(mysql.ClientConfig{DbName: config.MySQLPrimaryDBName}),
-			mysql.WithLocation(config.MySQLPrimaryHost, config.MySQLPrimaryPort),
-			mysql.ParseTime(),
+		dbPrimary, err := mysql_client.NewClient(
+			mysql_client.WithAuth(config.MySQLPrimaryUser, config.MySQLPrimaryPassword),
+			mysql_client.WithConfig(mysql_client.ClientConfig{DbName: config.MySQLPrimaryDBName}),
+			mysql_client.WithLocation(config.MySQLPrimaryHost, config.MySQLPrimaryPort),
+			mysql_client.ParseTime(),
 		)
 		if err != nil {
 			return nil, err
 		}
 
-		dbSecondary, err := mysql.NewClient(
-			mysql.WithAuth(config.MySQLSecondaryUser, config.MySQLSecondaryPassword),
-			mysql.WithConfig(mysql.ClientConfig{DbName: config.MySQLSecondaryDBName}),
-			mysql.WithLocation(config.MySQLSecondaryHost, config.MySQLSecondaryPort),
-			mysql.ParseTime(),
+		dbSecondary, err := mysql_client.NewClient(
+			mysql_client.WithAuth(config.MySQLSecondaryUser, config.MySQLSecondaryPassword),
+			mysql_client.WithConfig(mysql_client.ClientConfig{DbName: config.MySQLSecondaryDBName}),
+			mysql_client.WithLocation(config.MySQLSecondaryHost, config.MySQLSecondaryPort),
+			mysql_client.ParseTime(),
 		)
 		if err != nil {
 			return nil, err
@@ -64,8 +64,8 @@ func NewDefaultRepository(config *Config) (repository.Provider, error) {
 			return nil, err
 		}
 
-		repo, err = repository_mysql.NewRepository(
-			repository_mysql.WithGormClient(dbClient),
+		repo, err = mysql.NewRepository(
+			mysql.WithGormClient(dbClient),
 		)
 		if err != nil {
 			return nil, err
