+++ internal/repository/mysql/barrel_test.go
@@ -1,4 +1,4 @@
-package repository_mysql_test
+package mysql_test
 
 import (
 	"context"
@@ -6,29 +6,26 @@ import (
 	"fmt"
 	"regexp"
 	"strings"
-
 	"time"
 
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/go-seidon/chariot/internal/repository"
-	repository_mysql "github.com/go-seidon/chariot/internal/repository-mysql"
+	"github.com/go-seidon/chariot/internal/repository/mysql"
 	"github.com/go-seidon/provider/typeconv"
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	gorm_mysql "gorm.io/driver/mysql"
+	"gorm.io/gorm"
 )
 
-var _ = Describe("Auth Repository", func() {
-
-	Context("CreateClient function", Label("unit"), func() {
+var _ = Describe("Barrel Repository", func() {
+	Context("CreateBarrel function", Label("unit"), func() {
 		var (
 			ctx        context.Context
 			currentTs  time.Time
 			dbClient   sqlmock.Sqlmock
-			authRepo   repository.Auth
-			p          repository.CreateClientParam
+			barrelRepo repository.Barrel
+			p          repository.CreateBarrelParam
 			checkStmt  string
 			insertStmt string
 			findStmt   string
@@ -47,7 +44,7 @@ var _ = Describe("Auth Repository", func() {
 				AbortSuite("failed create db mock: " + err.Error())
 			}
 
-			gormClient, err := gorm.Open(mysql.New(mysql.Config{
+			gormClient, err := gorm.Open(gorm_mysql.New(gorm_mysql.Config{
 				Conn:                      db,
 				SkipInitializeWithVersion: true,
 			}), &gorm.Config{
@@ -56,22 +53,21 @@ var _ = Describe("Auth Repository", func() {
 			if err != nil {
 				AbortSuite("failed create gorm client: " + err.Error())
 			}
-			authRepo = repository_mysql.NewAuth(repository_mysql.AuthParam{
+			barrelRepo = mysql.NewBarrel(mysql.BarrelParam{
 				GormClient: gormClient,
 			})
 
-			p = repository.CreateClientParam{
-				Id:           "id",
-				ClientId:     "client-id",
-				ClientSecret: "client-secret",
-				Name:         "name",
-				Type:         "basic",
-				Status:       "active",
-				CreatedAt:    currentTs,
+			p = repository.CreateBarrelParam{
+				Id:        "id",
+				Code:      "code",
+				Name:      "name",
+				Provider:  "goseidon_hippo",
+				Status:    "active",
+				CreatedAt: currentTs,
 			}
-			checkStmt = regexp.QuoteMeta("SELECT id, client_id FROM `auth_client` WHERE client_id = ? ORDER BY `auth_client`.`id` LIMIT 1")
-			insertStmt = regexp.QuoteMeta("INSERT INTO `auth_client` (`id`,`client_id`,`client_secret`,`name`,`type`,`status`,`created_at`,`updated_at`) VALUES (?,?,?,?,?,?,?,?)")
-			findStmt = regexp.QuoteMeta("SELECT id, client_id, client_secret, name, type, status, created_at FROM `auth_client` WHERE id = ? ORDER BY `auth_client`.`id` LIMIT 1")
+			checkStmt = regexp.QuoteMeta("SELECT id, code FROM `barrel` WHERE code = ? ORDER BY `barrel`.`id` LIMIT 1")
+			insertStmt = regexp.QuoteMeta("INSERT INTO `barrel` (`id`,`code`,`name`,`provider`,`status`,`created_at`,`updated_at`) VALUES (?,?,?,?,?,?,?)")
+			findStmt = regexp.QuoteMeta("SELECT id, code, name, provider, status, created_at FROM `barrel` WHERE id = ? ORDER BY `barrel`.`id` LIMIT 1")
 		})
 
 		AfterEach(func() {
@@ -87,95 +83,95 @@ var _ = Describe("Auth Repository", func() {
 					ExpectBegin().
 					WillReturnError(fmt.Errorf("begin error"))
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("begin error")))
 			})
 		})
 
-		When("failed rollback during check client", func() {
+		When("failed rollback during check barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(fmt.Errorf("network error"))
 
 				dbClient.
 					ExpectRollback().
 					WillReturnError(fmt.Errorf("rollback error"))
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("rollback error")))
 			})
 		})
 
-		When("failed check client", func() {
+		When("failed check barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(fmt.Errorf("network error"))
 
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
 			})
 		})
 
-		When("client already exists", func() {
+		When("barrel already exists", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				rows := sqlmock.NewRows([]string{
-					"id", "client_id",
+					"id", "code",
 				}).AddRow(
-					p.Id, p.ClientId,
+					p.Id, p.Code,
 				)
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnRows(rows)
 
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(repository.ErrExists))
 			})
 		})
 
-		When("failed rollback during client creation", func() {
+		When("failed rollback during barrel creation", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(gorm.ErrRecordNotFound)
 
 				dbClient.
 					ExpectExec(insertStmt).
 					WithArgs(
-						p.Id, p.ClientId, p.ClientSecret,
-						p.Name, p.Type, p.Status,
+						p.Id, p.Code,
+						p.Name, p.Provider, p.Status,
 						p.CreatedAt.UnixMilli(),
 						p.CreatedAt.UnixMilli(),
 					).
@@ -185,28 +181,28 @@ var _ = Describe("Auth Repository", func() {
 					ExpectRollback().
 					WillReturnError(fmt.Errorf("rollback error"))
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("rollback error")))
 			})
 		})
 
-		When("failed create client", func() {
+		When("failed create barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(gorm.ErrRecordNotFound)
 
 				dbClient.
 					ExpectExec(insertStmt).
 					WithArgs(
-						p.Id, p.ClientId, p.ClientSecret,
-						p.Name, p.Type, p.Status,
+						p.Id, p.Code,
+						p.Name, p.Provider, p.Status,
 						p.CreatedAt.UnixMilli(),
 						p.CreatedAt.UnixMilli(),
 					).
@@ -215,28 +211,28 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
 			})
 		})
 
-		When("failed rollback during find client", func() {
+		When("failed rollback during find barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(gorm.ErrRecordNotFound)
 
 				dbClient.
 					ExpectExec(insertStmt).
 					WithArgs(
-						p.Id, p.ClientId, p.ClientSecret,
-						p.Name, p.Type, p.Status,
+						p.Id, p.Code,
+						p.Name, p.Provider, p.Status,
 						p.CreatedAt.UnixMilli(),
 						p.CreatedAt.UnixMilli(),
 					).
@@ -251,28 +247,28 @@ var _ = Describe("Auth Repository", func() {
 					ExpectRollback().
 					WillReturnError(fmt.Errorf("rollback error"))
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("rollback error")))
 			})
 		})
 
-		When("failed find client", func() {
+		When("failed find barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(gorm.ErrRecordNotFound)
 
 				dbClient.
 					ExpectExec(insertStmt).
 					WithArgs(
-						p.Id, p.ClientId, p.ClientSecret,
-						p.Name, p.Type, p.Status,
+						p.Id, p.Code,
+						p.Name, p.Provider, p.Status,
 						p.CreatedAt.UnixMilli(),
 						p.CreatedAt.UnixMilli(),
 					).
@@ -286,39 +282,41 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
 			})
 		})
 
-		When("failed commit during success create client", func() {
+		When("failed commit during success create barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(gorm.ErrRecordNotFound)
 
 				dbClient.
 					ExpectExec(insertStmt).
 					WithArgs(
-						p.Id, p.ClientId, p.ClientSecret,
-						p.Name, p.Type, p.Status,
+						p.Id, p.Code,
+						p.Name, p.Provider, p.Status,
 						p.CreatedAt.UnixMilli(),
 						p.CreatedAt.UnixMilli(),
 					).
 					WillReturnResult(sqlmock.NewResult(1, 1))
 
 				rows := sqlmock.NewRows([]string{
-					"id", "client_id", "client_secret",
-					"name", "type", "status", "created_at",
+					"id", "code",
+					"name", "provider",
+					"status", "created_at",
 				}).AddRow(
-					p.Id, p.ClientId, p.ClientSecret,
-					p.Name, p.Type, p.Status, p.CreatedAt.UnixMilli(),
+					p.Id, p.Code,
+					p.Name, p.Provider,
+					p.Status, p.CreatedAt.UnixMilli(),
 				)
 				dbClient.
 					ExpectQuery(findStmt).
@@ -329,39 +327,41 @@ var _ = Describe("Auth Repository", func() {
 					ExpectCommit().
 					WillReturnError(fmt.Errorf("commit error"))
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("commit error")))
 			})
 		})
 
-		When("success create client", func() {
+		When("success create barrel", func() {
 			It("should return result", func() {
 				dbClient.
 					ExpectBegin()
 
 				dbClient.
 					ExpectQuery(checkStmt).
-					WithArgs(p.ClientId).
+					WithArgs(p.Code).
 					WillReturnError(gorm.ErrRecordNotFound)
 
 				dbClient.
 					ExpectExec(insertStmt).
 					WithArgs(
-						p.Id, p.ClientId, p.ClientSecret,
-						p.Name, p.Type, p.Status,
+						p.Id, p.Code,
+						p.Name, p.Provider, p.Status,
 						p.CreatedAt.UnixMilli(),
 						p.CreatedAt.UnixMilli(),
 					).
 					WillReturnResult(sqlmock.NewResult(1, 1))
 
 				rows := sqlmock.NewRows([]string{
-					"id", "client_id", "client_secret",
-					"name", "type", "status", "created_at",
+					"id", "code",
+					"name", "provider",
+					"status", "created_at",
 				}).AddRow(
-					p.Id, p.ClientId, p.ClientSecret,
-					p.Name, p.Type, p.Status, p.CreatedAt.UnixMilli(),
+					p.Id, p.Code,
+					p.Name, p.Provider,
+					p.Status, p.CreatedAt.UnixMilli(),
 				)
 				dbClient.
 					ExpectQuery(findStmt).
@@ -371,16 +371,15 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectCommit()
 
-				res, err := authRepo.CreateClient(ctx, p)
+				res, err := barrelRepo.CreateBarrel(ctx, p)
 
-				expectedRes := &repository.CreateClientResult{
-					Id:           p.Id,
-					ClientId:     p.ClientId,
-					ClientSecret: p.ClientSecret,
-					Name:         p.Name,
-					Type:         p.Type,
-					Status:       p.Status,
-					CreatedAt:    time.UnixMilli(p.CreatedAt.UnixMilli()).UTC(),
+				expectedRes := &repository.CreateBarrelResult{
+					Id:        p.Id,
+					Code:      p.Code,
+					Name:      p.Name,
+					Provider:  p.Provider,
+					Status:    p.Status,
+					CreatedAt: time.UnixMilli(p.CreatedAt.UnixMilli()).UTC(),
 				}
 				Expect(res).To(Equal(expectedRes))
 				Expect(err).To(BeNil())
@@ -388,16 +387,16 @@ var _ = Describe("Auth Repository", func() {
 		})
 	})
 
-	Context("FindClient function", Label("unit"), func() {
+	Context("FindBarrel function", Label("unit"), func() {
 
 		var (
-			ctx       context.Context
-			currentTs time.Time
-			dbClient  sqlmock.Sqlmock
-			authRepo  repository.Auth
-			p         repository.FindClientParam
-			r         *repository.FindClientResult
-			findStmt  string
+			ctx        context.Context
+			currentTs  time.Time
+			dbClient   sqlmock.Sqlmock
+			barrelRepo repository.Barrel
+			p          repository.FindBarrelParam
+			r          *repository.FindBarrelResult
+			findStmt   string
 		)
 
 		BeforeEach(func() {
@@ -413,7 +412,7 @@ var _ = Describe("Auth Repository", func() {
 				AbortSuite("failed create db mock: " + err.Error())
 			}
 
-			gormClient, err := gorm.Open(mysql.New(mysql.Config{
+			gormClient, err := gorm.Open(gorm_mysql.New(gorm_mysql.Config{
 				Conn:                      db,
 				SkipInitializeWithVersion: true,
 			}), &gorm.Config{
@@ -422,23 +421,23 @@ var _ = Describe("Auth Repository", func() {
 			if err != nil {
 				AbortSuite("failed create gorm client: " + err.Error())
 			}
-			authRepo = repository_mysql.NewAuth(repository_mysql.AuthParam{
+			barrelRepo = mysql.NewBarrel(mysql.BarrelParam{
 				GormClient: gormClient,
 			})
 
-			p = repository.FindClientParam{
+			p = repository.FindBarrelParam{
 				Id: "id",
 			}
-			r = &repository.FindClientResult{
-				Id:           "id",
-				ClientId:     "client-id",
-				ClientSecret: "client-secret",
-				Name:         "name",
-				Type:         "basic",
-				Status:       "active",
-				CreatedAt:    time.UnixMilli(currentTs.UnixMilli()).UTC(),
+			r = &repository.FindBarrelResult{
+				Id:        "id",
+				Code:      "code",
+				Name:      "name",
+				Provider:  "goseidon_hippo",
+				Status:    "active",
+				CreatedAt: time.UnixMilli(currentTs.UnixMilli()).UTC(),
+				UpdatedAt: typeconv.Time(time.UnixMilli(currentTs.UnixMilli()).UTC()),
 			}
-			findStmt = regexp.QuoteMeta("SELECT id, client_id, client_secret, name, type, status, created_at, updated_at FROM `auth_client` WHERE id = ? ORDER BY `auth_client`.`id` LIMIT 1")
+			findStmt = regexp.QuoteMeta("SELECT id, code, name, provider, status, created_at, updated_at FROM `barrel` WHERE id = ? ORDER BY `barrel`.`id` LIMIT 1")
 		})
 
 		AfterEach(func() {
@@ -448,43 +447,43 @@ var _ = Describe("Auth Repository", func() {
 			}
 		})
 
-		When("failed check client", func() {
+		When("failed check barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectQuery(findStmt).
 					WithArgs(p.Id).
 					WillReturnError(fmt.Errorf("network error"))
 
-				res, err := authRepo.FindClient(ctx, p)
+				res, err := barrelRepo.FindBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
 			})
 		})
 
-		When("client is not available", func() {
+		When("barrel is not available", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectQuery(findStmt).
 					WithArgs(p.Id).
 					WillReturnError(gorm.ErrRecordNotFound)
 
-				res, err := authRepo.FindClient(ctx, p)
+				res, err := barrelRepo.FindBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(repository.ErrNotFound))
 			})
 		})
 
-		When("success find client", func() {
+		When("success find barrel", func() {
 			It("should return result", func() {
 				rows := sqlmock.NewRows([]string{
-					"id", "client_id", "client_secret",
-					"name", "type", "status",
+					"id", "code",
+					"name", "provider", "status",
 					"created_at", "updated_at",
 				}).AddRow(
-					r.Id, r.ClientId, r.ClientSecret,
-					r.Name, r.Type, r.Status,
+					r.Id, r.Code,
+					r.Name, r.Provider, r.Status,
 					currentTs.UnixMilli(), currentTs.UnixMilli(),
 				)
 
@@ -493,34 +492,23 @@ var _ = Describe("Auth Repository", func() {
 					WithArgs(p.Id).
 					WillReturnRows(rows)
 
-				res, err := authRepo.FindClient(ctx, p)
+				res, err := barrelRepo.FindBarrel(ctx, p)
 
-				updatedAt := time.UnixMilli(currentTs.UnixMilli()).UTC()
-				r := &repository.FindClientResult{
-					Id:           "id",
-					ClientId:     "client-id",
-					ClientSecret: "client-secret",
-					Name:         "name",
-					Type:         "basic",
-					Status:       "active",
-					CreatedAt:    time.UnixMilli(currentTs.UnixMilli()).UTC(),
-					UpdatedAt:    &updatedAt,
-				}
 				Expect(res).To(Equal(r))
 				Expect(err).To(BeNil())
 			})
 		})
 	})
 
-	Context("UpdateClient function", Label("unit"), func() {
+	Context("UpdateBarrel function", Label("unit"), func() {
 
 		var (
 			ctx        context.Context
 			currentTs  time.Time
 			dbClient   sqlmock.Sqlmock
-			authRepo   repository.Auth
-			p          repository.UpdateClientParam
-			r          *repository.UpdateClientResult
+			barrelRepo repository.Barrel
+			p          repository.UpdateBarrelParam
+			r          *repository.UpdateBarrelResult
 			findStmt   string
 			updateStmt string
 			checkStmt  string
@@ -541,7 +529,7 @@ var _ = Describe("Auth Repository", func() {
 				AbortSuite("failed create db mock: " + err.Error())
 			}
 
-			gormClient, err := gorm.Open(mysql.New(mysql.Config{
+			gormClient, err := gorm.Open(gorm_mysql.New(gorm_mysql.Config{
 				Conn:                      db,
 				SkipInitializeWithVersion: true,
 			}), &gorm.Config{
@@ -550,45 +538,45 @@ var _ = Describe("Auth Repository", func() {
 			if err != nil {
 				AbortSuite("failed create gorm client: " + err.Error())
 			}
-			authRepo = repository_mysql.NewAuth(repository_mysql.AuthParam{
+			barrelRepo = mysql.NewBarrel(mysql.BarrelParam{
 				GormClient: gormClient,
 			})
 
-			p = repository.UpdateClientParam{
+			p = repository.UpdateBarrelParam{
 				Id:        "id",
-				ClientId:  "new-client-id",
+				Code:      "new-code",
 				Name:      "new-name",
-				Type:      "basic",
+				Provider:  "goseidon_hippo",
 				Status:    "active",
 				UpdatedAt: currentTs,
 			}
-			r = &repository.UpdateClientResult{
-				Id:           "id",
-				ClientId:     "new-client-id",
-				ClientSecret: "client-secret",
-				Name:         "new-name",
-				Type:         "basic",
-				Status:       "active",
-				CreatedAt:    time.UnixMilli(currentTs.UnixMilli()).UTC(),
-				UpdatedAt:    time.UnixMilli(currentTs.UnixMilli()).UTC(),
+			r = &repository.UpdateBarrelResult{
+				Id:        "id",
+				Code:      "new-code",
+				Name:      "new-name",
+				Provider:  "goseidon_hippo",
+				Status:    "active",
+				CreatedAt: time.UnixMilli(currentTs.UnixMilli()).UTC(),
+				UpdatedAt: time.UnixMilli(currentTs.UnixMilli()).UTC(),
 			}
-			findStmt = regexp.QuoteMeta("SELECT id, client_id, name, type, status FROM `auth_client` WHERE id = ? ORDER BY `auth_client`.`id` LIMIT 1")
-			updateStmt = regexp.QuoteMeta("UPDATE `auth_client` SET `client_id`=?,`name`=?,`status`=?,`type`=?,`updated_at`=? WHERE id = ?")
-			checkStmt = regexp.QuoteMeta("SELECT id, client_id, client_secret, name, type, status, created_at, updated_at FROM `auth_client` WHERE id = ? ORDER BY `auth_client`.`id` LIMIT 1")
+			findStmt = regexp.QuoteMeta("SELECT id, code, name, provider, status FROM `barrel` WHERE id = ? ORDER BY `barrel`.`id` LIMIT 1")
+			updateStmt = regexp.QuoteMeta("UPDATE `barrel` SET `code`=?,`name`=?,`provider`=?,`status`=?,`updated_at`=? WHERE id = ?")
+			checkStmt = regexp.QuoteMeta("SELECT id, code, name, provider, status, created_at, updated_at FROM `barrel` WHERE id = ? ORDER BY `barrel`.`id` LIMIT 1")
 			findRows = sqlmock.NewRows([]string{
-				"id", "client_id",
-				"name", "type", "status",
+				"id", "code",
+				"name", "provider", "status",
 			}).AddRow(
-				"id", "old-client-id",
-				"old-name", "basic", "inactive",
+				"id", "old-code",
+				"old-name", "goseidon_hippo", "inactive",
 			)
 			checkRows = sqlmock.NewRows([]string{
-				"id", "client_id", "client_secret",
-				"name", "type", "status",
+				"id", "code",
+				"name", "provider",
+				"status",
 				"created_at", "updated_at",
 			}).AddRow(
-				"id", "new-client-id", "client-secret",
-				"new-name", "basic", "active",
+				"id", "new-code",
+				"new-name", "goseidon_hippo", "active",
 				currentTs.UnixMilli(), currentTs.UnixMilli(),
 			)
 		})
@@ -606,14 +594,14 @@ var _ = Describe("Auth Repository", func() {
 					ExpectBegin().
 					WillReturnError(fmt.Errorf("begin error"))
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("begin error")))
 			})
 		})
 
-		When("failed rollback during find client", func() {
+		When("failed rollback during find barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
@@ -627,14 +615,14 @@ var _ = Describe("Auth Repository", func() {
 					ExpectRollback().
 					WillReturnError(fmt.Errorf("rollback error"))
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("rollback error")))
 			})
 		})
 
-		When("failed find client", func() {
+		When("failed find barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
@@ -647,14 +635,14 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
 			})
 		})
 
-		When("client is not available", func() {
+		When("barrel is not available", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
@@ -667,14 +655,14 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(repository.ErrNotFound))
 			})
 		})
 
-		When("failed rollback during update client", func() {
+		When("failed rollback during update barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
@@ -687,10 +675,10 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectExec(updateStmt).
 					WithArgs(
-						p.ClientId,
+						p.Code,
 						p.Name,
+						p.Provider,
 						p.Status,
-						p.Type,
 						p.UpdatedAt.UnixMilli(),
 						p.Id,
 					).
@@ -700,14 +688,14 @@ var _ = Describe("Auth Repository", func() {
 					ExpectRollback().
 					WillReturnError(fmt.Errorf("rollback error"))
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("rollback error")))
 			})
 		})
 
-		When("failed update client", func() {
+		When("failed update barrel", func() {
 			It("should return error", func() {
 				dbClient.
 					ExpectBegin()
@@ -720,10 +708,10 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectExec(updateStmt).
 					WithArgs(
-						p.ClientId,
+						p.Code,
 						p.Name,
+						p.Provider,
 						p.Status,
-						p.Type,
 						p.UpdatedAt.UnixMilli(),
 						p.Id,
 					).
@@ -732,7 +720,7 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
@@ -752,10 +740,10 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectExec(updateStmt).
 					WithArgs(
-						p.ClientId,
+						p.Code,
 						p.Name,
+						p.Provider,
 						p.Status,
-						p.Type,
 						p.UpdatedAt.UnixMilli(),
 						p.Id,
 					).
@@ -770,7 +758,7 @@ var _ = Describe("Auth Repository", func() {
 					ExpectRollback().
 					WillReturnError(fmt.Errorf("rollback error"))
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("rollback error")))
@@ -790,10 +778,10 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectExec(updateStmt).
 					WithArgs(
-						p.ClientId,
+						p.Code,
 						p.Name,
+						p.Provider,
 						p.Status,
-						p.Type,
 						p.UpdatedAt.UnixMilli(),
 						p.Id,
 					).
@@ -807,7 +795,7 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectRollback()
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
@@ -827,10 +815,10 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectExec(updateStmt).
 					WithArgs(
-						p.ClientId,
+						p.Code,
 						p.Name,
+						p.Provider,
 						p.Status,
-						p.Type,
 						p.UpdatedAt.UnixMilli(),
 						p.Id,
 					).
@@ -845,14 +833,14 @@ var _ = Describe("Auth Repository", func() {
 					ExpectCommit().
 					WillReturnError(fmt.Errorf("commit error"))
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("commit error")))
 			})
 		})
 
-		When("success update client", func() {
+		When("success update barrel", func() {
 			It("should return result", func() {
 				dbClient.
 					ExpectBegin()
@@ -865,10 +853,10 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectExec(updateStmt).
 					WithArgs(
-						p.ClientId,
+						p.Code,
 						p.Name,
+						p.Provider,
 						p.Status,
-						p.Type,
 						p.UpdatedAt.UnixMilli(),
 						p.Id,
 					).
@@ -882,7 +870,7 @@ var _ = Describe("Auth Repository", func() {
 				dbClient.
 					ExpectCommit()
 
-				res, err := authRepo.UpdateClient(ctx, p)
+				res, err := barrelRepo.UpdateBarrel(ctx, p)
 
 				Expect(res).To(Equal(r))
 				Expect(err).To(BeNil())
@@ -890,15 +878,15 @@ var _ = Describe("Auth Repository", func() {
 		})
 	})
 
-	Context("SearchClient function", Label("unit"), func() {
+	Context("SearchBarrel function", Label("unit"), func() {
 
 		var (
 			ctx        context.Context
 			currentTs  time.Time
 			dbClient   sqlmock.Sqlmock
-			authRepo   repository.Auth
-			p          repository.SearchClientParam
-			r          *repository.SearchClientResult
+			barrelRepo repository.Barrel
+			p          repository.SearchBarrelParam
+			r          *repository.SearchBarrelResult
 			searchStmt string
 			countStmt  string
 			searchRows *sqlmock.Rows
@@ -918,7 +906,7 @@ var _ = Describe("Auth Repository", func() {
 				AbortSuite("failed create db mock: " + err.Error())
 			}
 
-			gormClient, err := gorm.Open(mysql.New(mysql.Config{
+			gormClient, err := gorm.Open(gorm_mysql.New(gorm_mysql.Config{
 				Conn:                      db,
 				SkipInitializeWithVersion: true,
 			}), &gorm.Config{
@@ -927,69 +915,72 @@ var _ = Describe("Auth Repository", func() {
 			if err != nil {
 				AbortSuite("failed create gorm client: " + err.Error())
 			}
-			authRepo = repository_mysql.NewAuth(repository_mysql.AuthParam{
+			barrelRepo = mysql.NewBarrel(mysql.BarrelParam{
 				GormClient: gormClient,
 			})
 
-			p = repository.SearchClientParam{
-				Limit:    24,
-				Offset:   48,
-				Keyword:  "goseidon",
-				Statuses: []string{"active"},
+			p = repository.SearchBarrelParam{
+				Limit:     24,
+				Offset:    48,
+				Keyword:   "goseidon",
+				Statuses:  []string{"active"},
+				Providers: []string{"goseidon_hippo"},
+				Codes:     []string{"hippo1"},
 			}
-			updatedAt := time.UnixMilli(currentTs.UnixMilli()).UTC()
-			r = &repository.SearchClientResult{
-				Summary: repository.SearchClientSummary{
+			r = &repository.SearchBarrelResult{
+				Summary: repository.SearchBarrelSummary{
 					TotalItems: 2,
 				},
-				Items: []repository.SearchClientItem{
+				Items: []repository.SearchBarrelItem{
 					{
-						Id:           "id-1",
-						ClientId:     "client-id-1",
-						ClientSecret: "client-secret-1",
-						Name:         "name-1",
-						Type:         "basic",
-						Status:       "inactive",
-						CreatedAt:    time.UnixMilli(currentTs.UnixMilli()).UTC(),
-						UpdatedAt:    &updatedAt,
+						Id:        "id-1",
+						Code:      "code-1",
+						Name:      "name-1",
+						Provider:  "goseidon_hippo",
+						Status:    "inactive",
+						CreatedAt: time.UnixMilli(currentTs.UnixMilli()).UTC(),
+						UpdatedAt: typeconv.Time(time.UnixMilli(currentTs.UnixMilli()).UTC()),
 					},
 					{
-						Id:           "id-2",
-						ClientId:     "client-id-2",
-						ClientSecret: "client-secret-2",
-						Name:         "name-2",
-						Type:         "basic",
-						Status:       "active",
-						CreatedAt:    time.UnixMilli(currentTs.UnixMilli()).UTC(),
-						UpdatedAt:    &updatedAt,
+						Id:        "id-2",
+						Code:      "code-2",
+						Name:      "name-2",
+						Provider:  "goseidon_hippo",
+						Status:    "active",
+						CreatedAt: time.UnixMilli(currentTs.UnixMilli()).UTC(),
+						UpdatedAt: typeconv.Time(time.UnixMilli(currentTs.UnixMilli()).UTC()),
 					},
 				},
 			}
 			searchStmt = regexp.QuoteMeta(strings.TrimSpace(`
-				SELECT id, client_id, client_secret, name, type, status, created_at, updated_at 
-				FROM ` + "`auth_client`" + ` 
-				WHERE name LIKE ? OR client_id LIKE ?
+				SELECT id, code, name, provider, status, created_at, updated_at 
+				FROM ` + "`barrel`" + ` 
+				WHERE name LIKE ? OR code LIKE ?
 				AND status IN (?)
+				AND provider IN (?)
+				AND code IN (?)
 				LIMIT 24
 				OFFSET 48
 			`))
 			countStmt = regexp.QuoteMeta(strings.TrimSpace(`
 				SELECT count(*)
-				FROM ` + "`auth_client`" + ` 
-				WHERE name LIKE ? OR client_id LIKE ?
+				FROM ` + "`barrel`" + ` 
+				WHERE name LIKE ? OR code LIKE ?
 				AND status IN (?)
+				AND provider IN (?)
+				AND code IN (?)
 			`))
 			searchRows = sqlmock.NewRows([]string{
-				"id", "client_id", "client_secret",
-				"name", "type", "status",
+				"id", "code",
+				"name", "provider", "status",
 				"created_at", "updated_at",
 			}).AddRow(
-				"id-1", "client-id-1", "client-secret-1",
-				"name-1", "basic", "inactive",
+				"id-1", "code-1",
+				"name-1", "goseidon_hippo", "inactive",
 				currentTs.UnixMilli(), currentTs.UnixMilli(),
 			).AddRow(
-				"id-2", "client-id-2", "client-secret-2",
-				"name-2", "basic", "active",
+				"id-2", "code-2",
+				"name-2", "goseidon_hippo", "active",
 				currentTs.UnixMilli(), currentTs.UnixMilli(),
 			)
 			countRows = sqlmock.
@@ -1012,10 +1003,12 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnError(fmt.Errorf("network error"))
 
-				res, err := authRepo.SearchClient(ctx, p)
+				res, err := barrelRepo.SearchBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
@@ -1030,16 +1023,18 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnError(gorm.ErrRecordNotFound)
 
-				res, err := authRepo.SearchClient(ctx, p)
+				res, err := barrelRepo.SearchBarrel(ctx, p)
 
-				r := &repository.SearchClientResult{
-					Summary: repository.SearchClientSummary{
+				r := &repository.SearchBarrelResult{
+					Summary: repository.SearchBarrelSummary{
 						TotalItems: 0,
 					},
-					Items: []repository.SearchClientItem{},
+					Items: []repository.SearchBarrelItem{},
 				}
 				Expect(res).To(Equal(r))
 				Expect(err).To(BeNil())
@@ -1049,13 +1044,13 @@ var _ = Describe("Auth Repository", func() {
 		When("failed count search client", func() {
 			It("should return result", func() {
 				searchRows := sqlmock.NewRows([]string{
-					"id", "client_id", "client_secret",
-					"name", "type", "status",
+					"id", "code",
+					"name", "provider", "status",
 					"created_at", "updated_at",
 				}).AddRow(
-					"id-1", "client-id-1", "client-secret-1",
-					"name-1", "basic", "inactive",
-					currentTs.UnixMilli(), nil,
+					"id-1", "code-1",
+					"name-1", "goseidon_hippo", "inactive",
+					currentTs.UnixMilli(), currentTs.UnixMilli(),
 				)
 				dbClient.
 					ExpectQuery(searchStmt).
@@ -1063,6 +1058,8 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnRows(searchRows)
 
@@ -1072,10 +1069,12 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnError(fmt.Errorf("network error"))
 
-				res, err := authRepo.SearchClient(ctx, p)
+				res, err := barrelRepo.SearchBarrel(ctx, p)
 
 				Expect(res).To(BeNil())
 				Expect(err).To(Equal(fmt.Errorf("network error")))
@@ -1085,12 +1084,12 @@ var _ = Describe("Auth Repository", func() {
 		When("there is one client", func() {
 			It("should return result", func() {
 				searchRows := sqlmock.NewRows([]string{
-					"id", "client_id", "client_secret",
-					"name", "type", "status",
+					"id", "code",
+					"name", "provider", "status",
 					"created_at", "updated_at",
 				}).AddRow(
-					"id-1", "client-id-1", "client-secret-1",
-					"name-1", "basic", "inactive",
+					"id-1", "code-1",
+					"name-1", "goseidon_hippo", "inactive",
 					currentTs.UnixMilli(), currentTs.UnixMilli(),
 				)
 				dbClient.
@@ -1099,6 +1098,8 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnRows(searchRows)
 
@@ -1111,25 +1112,26 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnRows(countRows)
 
-				res, err := authRepo.SearchClient(ctx, p)
+				res, err := barrelRepo.SearchBarrel(ctx, p)
 
-				r := &repository.SearchClientResult{
-					Summary: repository.SearchClientSummary{
+				r := &repository.SearchBarrelResult{
+					Summary: repository.SearchBarrelSummary{
 						TotalItems: 1,
 					},
-					Items: []repository.SearchClientItem{
+					Items: []repository.SearchBarrelItem{
 						{
-							Id:           "id-1",
-							ClientId:     "client-id-1",
-							ClientSecret: "client-secret-1",
-							Name:         "name-1",
-							Type:         "basic",
-							Status:       "inactive",
-							CreatedAt:    time.UnixMilli(currentTs.UnixMilli()).UTC(),
-							UpdatedAt:    typeconv.Time(time.UnixMilli(currentTs.UnixMilli()).UTC()),
+							Id:        "id-1",
+							Code:      "code-1",
+							Name:      "name-1",
+							Provider:  "goseidon_hippo",
+							Status:    "inactive",
+							CreatedAt: time.UnixMilli(currentTs.UnixMilli()).UTC(),
+							UpdatedAt: typeconv.Time(time.UnixMilli(currentTs.UnixMilli()).UTC()),
 						},
 					},
 				}
@@ -1146,6 +1148,8 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnRows(searchRows)
 
@@ -1155,15 +1159,16 @@ var _ = Describe("Auth Repository", func() {
 						"%"+p.Keyword+"%",
 						"%"+p.Keyword+"%",
 						p.Statuses[0],
+						p.Providers[0],
+						p.Codes[0],
 					).
 					WillReturnRows(countRows)
 
-				res, err := authRepo.SearchClient(ctx, p)
+				res, err := barrelRepo.SearchBarrel(ctx, p)
 
 				Expect(res).To(Equal(r))
 				Expect(err).To(BeNil())
 			})
 		})
 	})
-
 })
